// src/types/heritageObj/HeritageObjType.ts
export type HeritageObjType =
  // official heritage
  | "architecture"
  | "history"
  | "history_monumental"
  | "monument"
  | "memorial"
  | "archaeology"
  | "garden_park"
  | "urban_planning"
  | "landscape"

  // local / urban elements
  | "local_place"
  | "local_monument"
  | "local_sculpture"
  | "local_art"
  | "local_element"
  | "viewpoint";